+++ src/index.js
@@ -1,4 +1,5 @@
 export * from './isPromise'
+export * from './stopEvent'
 export * from './useDebounce'
 export * from './useEventListener'
 export * from './useInput'