+++ src/index.js
@@ -1,3 +1,4 @@
+export * from './copyToClipboard'
 export * from './isPromise'
 export * from './stopEvent'
 export * from './useDebounce'