+++ CHANGELOG.md
@@ -19,9 +19,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### ✨ Enhancements
 
-- Create `useInterval` hook.
 - Create `isPromise` util.
 - Create `stopEvent` util.
+- Create `useInterval` hook.
+- Create `useTimeLeft` hook.
 
 ## 1.3.0 (June 11, 2020)