+++ CHANGELOG.md
@@ -15,11 +15,11 @@ All notable changes to this project will be documented here.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.0.0 (June 22, 2020)
 
 ### 🚨 Breaking changes
 
-- `useEventListener` : element paramater has been replaced by options. Set the targeted element into options object parameter instead.
+- `useEventListener` : element parameter has been replaced by options. Set the targeted element into options object parameter instead.
 
 ```diff
   useEventListener(